Make post-deploy-configure-k8s.sh idempotent
When creating loggregator-tls-cert-secrets, we previously failed if a
secret with that name already existed. This change will update the
secret idempotently.

This commit also adds a volume and volume mount to the
loggregator-fluentd daemonset pod spec to allow logs to be read when the
logs director is symlinked in.

The runner of the script can also now override the BOSH_ENV_NAME
variable if the given jq command doesn't work for their environment
Signed-off-by: Jen Spinney <jspinney@pivotal.io>