Fix port in auto-generated Eirini address
If the `opi.eirini_address` property is not explicitly specified, we use the IP of the current BOSH VM. We previously made a change to have this auto-generated address use `https` instead of `http`. In doing so, we accidentally appended the literal string "opi.tls_port" instead of the value of `opi.tls_port`.

This commit fixes that mistake.